refactor: rename signin files to kebab-case
- Rename normalForm.tsx to normal-form.tsx
- Rename oneMoreStep.tsx to one-more-step.tsx
- Update corresponding import paths in page.tsx
- Follow project naming conventions consistently
